06.01 log4j included

+++ src/main/java/by/my/web/EventController.java
@@ -9,8 +9,11 @@ import java.util.Date;
 import java.util.List;
 import java.util.Map;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.hibernate.HibernateException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -40,10 +43,16 @@ public class EventController {
 	@Autowired
 	MessageService messageService;
 
+	private static final Logger logger = LogManager
+			.getLogger(EventController.class);
+
 	@RequestMapping(value = "newEvent.html", method = RequestMethod.GET)
-	public String viewRegistration(Map<String, Object> model) {
+	public String viewRegistration(Map<String, Object> model,
+			HttpServletRequest request) {
 		Event event = new Event();
 		model.put("event", event);
+		logger.info("User:" + request.getUserPrincipal().getName()
+				+ "going to create new event.");
 		return "event/newEvent";
 	}
 
@@ -74,6 +83,8 @@ public class EventController {
 		}
 		eventService.createEvent(event, file);
 		session.setAttribute("imagesLoded", imageLoaded);
+		logger.info("New event:" + event.getEventName() + " was created by: "
+				+ event.getCreatedBy().getUsername());
 		return "redirect:/main.html";
 	}
 
@@ -84,13 +95,19 @@ public class EventController {
 		User joinedUser = userService.getUser(joinedUserName);
 		long userId = joinedUser.getId();
 		eventService.joinEvent(eventId, userId);
+		logger.info("Joined event ID: " + eventId + " by USER: " + joinedUserName);
 		return "redirect:/main.html";
 	}
 
 	@RequestMapping(value = "{eventId}/deleteEvent.html")
-	public String delete(@PathVariable long eventId, Model model) {
+	public String delete(@PathVariable long eventId, Model model,
+			Principal principal) {
 		Event event = eventService.getEventByID(eventId);
-		eventService.removeEvent(event);
+		if ((principal.getName().equals(event.getCreatedBy().getUsername()))
+				|| principal.getName().equals("Admin")) {
+			eventService.removeEvent(event);
+			logger.info("Event delited: " + event.getEventName() + " by user: " + principal.getName());
+		}
 		return "redirect:/main.html";
 	}
 
@@ -103,6 +120,7 @@ public class EventController {
 		long userId = joinedUser.getId();
 		eventService.unjoinEvent(eventId, userId);
 		model.addAttribute("event", event);
+		logger.info("Unjoined event ID: " + eventId + " by USER: " + joinedUserName);
 		return "redirect:/main.html";
 	}
 
@@ -111,9 +129,9 @@ public class EventController {
 			Model model) {
 		Event event = eventService.getEventByID(eventId);
 		List<Message> messageList = messageService.getEventMessages(event);
-		
 		model.addAttribute("event", event);
 		model.addAttribute("messages", messageList);
+		logger.info("Event details: " + event.getEventName());
 		return "event/eventDetails";
 	}
 
@@ -131,12 +149,10 @@ public class EventController {
 		message.setText(text);
 		message.setEvent(event);
 		messageService.save(message);
-
 		List<Message> messageList = messageService.getEventMessages(event);
-		System.out.println(messageList);
 		model.addAttribute("event", event);
 		model.addAttribute("messages", messageList);
-
+		logger.info("Post message: " + text + "; to event " + event.getEventName() + " ; by user " + user.getUsername());
 		return "event/eventDetails";
 	}
 }
